+++ R/utils.R
@@ -1,17 +1,34 @@
-workdata <- function(x, envir = parent.frame()){
-    ## 'x' is ignored currently
-    x <- c("ap", "ap7to9", "pcfr", "pcfr2to4")
+## TODO: remove workdata() later
+pcts_exdata <- function(x, envir = parent.frame()){
+    allnames <- c("ap", "ap7to9", "pcfr", "pcfr2to4")
+    if(missing(x))
+        x <- allnames
+    else if(isNA(x))
+        return(allnames)
+    else
+        stopifnot(is.character(x), length(x) > 0, all(x %in% allnames))
+
+    wrk <- vector("list", length(allnames))
+    names(wrk) <- allnames
     
-    ap <- pcts(datasets::AirPassengers)
-    ap7to9 <- window(ap, seasons = 7:9)
+    if(any(c("ap", "ap7to9") %in% x)){
+        wrk$ap <- pcts(datasets::AirPassengers)
+        if("ap7to9" %in% x)
+            wrk$ap7to9 <- window(wrk$ap, seasons = 7:9)
+    }
 
     ## without '::' there is a note from R CMD check
     ##     no visible binding for global variable dataFranses1996
-    pcfr <- pcts(pcts::dataFranses1996)
-    pcfr2to4 <- pcfr[2:4]
+    if(any(c("pcfr", "pcfr2to4") %in% x)){
+        wrk$pcfr <- pcts(pcts::dataFranses1996)
+        if("pcfr2to4" %in% x)
+            wrk$pcfr2to4 <- wrk$pcfr[2:4]
+    }
 
-    list2env(list(ap = ap, ap7to9 = ap7to9, pcfr = pcfr, pcfr2to4 = pcfr2to4), 
-             envir = envir)
+    ## retain only requested objects
+    wrk <- wrk[x]
+
+    list2env(wrk, envir = envir)
     x
 }